Task 2 - Employee Hourly Wage Calculation
Declare the function to calculate the hourly wage with the parameters of salary and hours per week. Calculate the hourly wage using the parameters plugged into the formula. Return the resulting hourly wage value to where it is called. Call the function calculateHourlyWage, plug in the values into the parameters, and log the wage to the console.